New Advancing System, 1st level map complete
Level 1 is now fully playable!!!

Fish now advance onto screen a set amount of time before their first note. This way multiple fish can be displayed on screen at once. This results in some clipping at the moment, but that will be fixed in time. Fish positions are also tentative.

On top of that, MusicManager.cs no longer directly communicates with FishHandler.cs, it only communicates with the base class RhythmVisuals.cs. This is so that future levels can use MusicManager.cs as is without any changes while having a different visuals system. Changes may be necessary later for this to be fully possible.